Rename team comm
