+++ .gitignore
@@ -221,5 +221,7 @@ doku/*
 /lib/
 
 # Path to the protocol buffer definitions, which are a diffrerent repository and managed by vcstool
-/humanoid_league_misc/humanoid_league_team_communication/humanoid_league_team_communication/RobocupProtocol
+/bitbots_team_communication/bitbots_team_communication/RobocupProtocol
+# Protobuf generated file
+/bitbots_team_communication/bitbots_team_communication/robocup_extension_pb2.py
