+++ bitbots_misc/bitbots_containers/hlvs/Dockerfile
@@ -61,7 +61,7 @@ ADD https://www.timeapi.io/api/Time/current/zone?timeZone=UTC /tmp/build-time
 RUN cd src/bitbots_main && \
     make pull-all && \
     rm -rf lib/udp_bridge bitbots_misc/bitbots_containers \
-    humanoid_league_visualization dynamic_stack_decider/dynamic_stack_decider_visualization bitbots_lowlevel \
+    lib/dynamic_stack_decider/dynamic_stack_decider_visualization bitbots_lowlevel \
     bitbots_wolfgang/wolfgang_pybullet_sim lib/DynamixelSDK lib/dynamixel-workbench \
     bitbots_misc/bitbots_basler_camera lib/pylon-ros-camera && \
     sed -i '/plotjuggler/d' bitbots_motion/bitbots_quintic_walk/package.xml && \
