+++ bitbots_team_communication/docs/_static/logo.png
