+++ bitbots_team_communication/launch/team_comm_standalone.launch
@@ -6,8 +6,8 @@
     <!-- load the global parameters -->
     <include file="$(find-pkg-share bitbots_utils)/launch/parameter_blackboard.launch" />
 
-    <node pkg="humanoid_league_team_communication" exec="team_comm.py" output="screen">
-        <param from="$(find-pkg-share humanoid_league_team_communication)/config/team_communication_config.yaml"/>
+    <node pkg="bitbots_team_communication" exec="team_comm.py" output="screen">
+        <param from="$(find-pkg-share bitbots_team_communication)/config/team_communication_config.yaml"/>
         <param name="use_sim_time" value="$(var sim)" />
     </node>
 </launch>
