+++ bitbots_team_communication/launch/team_comm_test_marker.launch
@@ -1,0 +1,11 @@
+<launch>
+    <arg name="rviz" default="true"/>
+
+    <group if="$(var rviz)">
+        <node pkg="rviz2" exec="rviz2" output="screen" args="-d $(find-pkg-share bitbots_team_communication)/config/team_comm_marker.rviz">
+        </node>
+    </group>
+    <node name="team_comm_test_marker" pkg="bitbots_team_communication" exec="team_comm_test_marker.py" output="screen">
+        <param from="$(find-pkg-share bitbots_team_communication)/config/team_communication_config.yaml" />
+    </node>
+</launch>
