+++ bitbots_team_communication/scripts/team_comm.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+import sys
+
+from bitbots_team_communication.bitbots_team_communication import main
+
+sys.exit(main())
