+++ bitbots_team_communication/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/bitbots_team_communication
+[install]
+install_scripts=$base/lib/bitbots_team_communication
