+++ bitbots_team_communication/setup.py
@@ -1,6 +1,6 @@
 from setuptools import find_packages, setup
 
-package_name = "humanoid_league_team_communication"
+package_name = "bitbots_team_communication"
 
 setup(
     name=package_name,
