--- humanoid_league_misc/humanoid_league_team_communication/.gitignore
@@ -1,1 +1,0 @@
-humanoid_league_team_communication/robocup_extension_pb2.py
