--- humanoid_league_misc/humanoid_league_team_communication/scripts/team_comm.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python3
-
-import sys
-
-from humanoid_league_team_communication.humanoid_league_team_communication import main
-
-sys.exit(main())
