--- humanoid_league_misc/humanoid_league_team_communication/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/humanoid_league_team_communication
-[install]
-install_scripts=$base/lib/humanoid_league_team_communication
